requirements.txt
fastapi==0.95.0
uvicorn==0.21.1
requests==2.28.2
python-decouple==3.8
mangum==0.17.0